requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
gunicorn==22.0.0
orjson==3.10.7
python-dateutil==2.9.0.post0
PyYAML==6.0.2
feedparser==6.0.11
bleach==6.1.0
requests==2.32.3
